Standardized text literal rules, moved shared vector rule names into implementation file.
