+++ cpp/src/nodes/node_list.cpp
@@ -4,7 +4,8 @@ NodeList::NodeList(NodePtrs elements, std::string delimiter) : elements(elements
 
 std::string NodeList::to_html(Json &context) {
   std::vector<std::string> html_vector;
-  std::transform(elements.begin(), elements.end(), html_vector.begin(), [&](NodePtr p) { return p->to_html(context); });
+  std::transform(elements.begin(), elements.end(), html_vector.begin(),
+    [&](NodePtr element) { return element->to_html(context); });
 
   return boost::algorithm::join(html_vector, delimiter);
 }
